Improve error handling and analytics
